Finished the lobby system
Handled leaving/disconnecting players from a room, even if it was the host and the leaving socket is now removed from the room (Host is passed to the next player).
Added a connected players row at the bottom of the room.
Refactored the way I was handling the slideOutAndChange function because it was calling functions that would define event listeners everytime they were called, resulting in stacked events (and everything being called *2 everytime slideOutAndChange was called).